+++ packages/core/avatar/src/helpers/index.js
@@ -1,4 +1,5 @@
 // @flow
+
 export { default as getPresenceSVG } from './getPresenceSVG';
 export { default as getProps } from './getProps';
 export { default as getStatusSVG } from './getStatusSVG';