+++ packages/core/avatar/src/styled/Icon.js
@@ -1,10 +1,10 @@
 // @flow
 import styled from 'styled-components';
-import { colors } from '@atlaskit/theme';
+import { colors, withTheme } from '@atlaskit/theme';
 import { BORDER_WIDTH } from './constants';
 
 // set fallbacks for border color/width to protect consumers from invalid values
-export const Outer = styled.span`
+export const Outer = withTheme(styled.span`
   align-content: center;
   align-items: center;
   background-color: ${props => props.bgColor || colors.background};
@@ -15,7 +15,7 @@ export const Outer = styled.span`
   overflow: hidden;
   padding: ${({ size }) => BORDER_WIDTH[size] || BORDER_WIDTH.medium}px;
   width: 100%;
-`;
+`);
 
 export const Inner = styled.span`
   align-items: center;