+++ packages/core/navigation-next/src/components/LayoutManagerWithViewController/types.js
@@ -5,19 +5,24 @@ import type { WithAnalyticsEventsProps } from '@atlaskit/analytics-next';
 
 import { UIController, ViewController } from '../../';
 
-import type { CollapseListeners, GetRefs } from '../LayoutManager/types';
+import type {
+  CollapseListeners,
+  ExperimentalFeatureFlags,
+  GetRefs,
+} from '../LayoutManager/types';
 
 import type { ViewControllerState } from '../../view-controller/types';
 
-export type LayoutManagerWithViewControllerProps = CollapseListeners & {
-  children: Node,
-  customComponents: { [string]: ComponentType<*> },
-  getRefs?: GetRefs,
-  globalNavigation: ComponentType<{}>,
-  navigationUIController: UIController,
-  navigationViewController: ViewController,
-  firstSkeletonToRender?: 'product' | 'container',
-};
+export type LayoutManagerWithViewControllerProps = CollapseListeners &
+  ExperimentalFeatureFlags & {
+    children: Node,
+    customComponents: { [string]: ComponentType<*> },
+    getRefs?: GetRefs,
+    globalNavigation: ComponentType<{}>,
+    navigationUIController: UIController,
+    navigationViewController: ViewController,
+    firstSkeletonToRender?: 'product' | 'container',
+  };
 
 export type LayoutManagerWithViewControllerState = {
   hasInitialised: boolean,