+++ packages/core/theme/src/components/__tests__/unit/Reset.js
@@ -2,8 +2,8 @@
 
 import React from 'react';
 import { mount } from 'enzyme';
-import Reset from '../Reset';
-import Theme from '../Theme';
+import Reset from '../../Reset';
+import Theme from '../../Theme';
 
 test('reset', () => {
   expect(mount(<Reset />)).toMatchSnapshot();
@@ -12,7 +12,13 @@ test('reset', () => {
 test('themed reset', () => {
   expect(
     mount(
-      <Theme backgroundColor="#000" textColor="#fff" linkColor="blue">
+      <Theme
+        theme={() => ({
+          backgroundColor: '#000',
+          textColor: '#fff',
+          linkColor: 'blue',
+        })}
+      >
         <Reset />
       </Theme>,
     ),