+++ packages/editor/editor-core/src/__tests__/integration/media/layout-mediaSingle.ts
@@ -13,7 +13,7 @@ BrowserTestCase(
   'Can change media single to full-width layout on fullpage',
   { skip: ['edge', 'ie', 'safari'] },
   async client => {
-    const browser = await new Page(client);
+    const browser = new Page(client);
 
     await browser.goto(fullpage.path);
     await browser.waitForSelector(editable);